requirements.txt
elasticsearch-dsl
Flask
Flask-Login
requests
pymongo
PyYAML
bcrypt